services/tasks_api/pyproject.toml
[tool.poetry]
name = "tasks-api"
version = "0.1.0"
description = ""
authors = ["Antonio Vilchez <antoniovmonge@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
httpx = "^0.27.0"
mangum = "^0.17.0"
boto3 = "^1.34.121"
pyjwt = "^2.8.0"
pydantic-settings = "^2.3.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.0.0"
bandit = "^1.7.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
